WebSockets/tests/submissions/Opera/binary/005.html

<!doctype html>
<title>WebSockets: Send/Receive ArrayBuffer, size less than network array buffer</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>
test(function(){
  var ws = new WebSocket(SCHEME_AND_DOMAIN+':8007/echo');
  var data = "";
  var datasize = 10;  
  ws.onopen = function(e) {
   ws.binaryType = "arraybuffer";
   data = new ArrayBuffer(datasize);
   ws.send(data);
  }
  ws.onmessage = function(e) {
    assert_equals(e.data.byteLength, datasize);
    end();
  }
 });
</script>
